# server/docs/swagger.json
{ "openapi":"3.0.0",
    "info":{ 
        "title":"CAB230 : Web Computing API",
        "description":"N10203036 Swagger Docs"
    },
    "servers":[ 
        {
            "url": "http://localhost"
        }
    ],
    "basePath":"/",
...
    "paths":{ 
        "/register":{ 
            "post":{ 
                "tags":["Authentication"],
                "summary": "Registers a new user account",
                "operationId":"register",
                "produces":["application/json"],
                "requestBody":{
                    "description":"required",
                    "required":"true",
                    "content":{
                        "application/json":{
                            "schema": {
                                "type": "object",                        
                                "properties":{
                                    "email":{
                                        "type":"string",
                                        "required":"true",
                                        "example":"demo@email.com"
                                    },
                                    "password":{
                                        "type":"string",
                                        "required":"true", 
                                        "example":"demoPassword"
                                    }
                                }          
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content":{
                            "application/json":{
                                "schema": {
                                    "type": "string",
                                    "properties":{
                                        "message":{
                                            "type":"string",
                                            "example":"yay! you've successfully registered your user account :)"
                                        }
                                    }
                                },
                                "example":{
                                    "message": "yay! you've successfully registered your user account :)"
                                }
                            }
                        }                   
                    },
                    "400": {
                        "description": "Bad Request",
                        "content":{
                            "application/json":{
                                "schema": {
                                    "type": "string",
                                    "properties":{
                                        "message":{
                                            "type":"string",
                                            "example":"oops! It looks like that user already exists :("
                                        }
                                    }
                                },
                                "example":{
                                    "message": "oops! It looks like that user already exists :("
                                  }
                            }
                        }
                    }
                }       
            }
        },
        "/login":{ 
            "post":{ 
                "tags":["Authentication"],
                "summary": "Login with an existing user account",
                "operationId":"login",
                "produces":["application/json"],
                "requestBody":{
                    "description":"required",
                    "required":"true",
                    "content":{
                        "application/json":{
                            "schema": {
                                "type": "object",                        
                                "properties":{
                                    "email":{
                                        "type":"string",
                                        "required":"true",
                                        "example":"demo@email.com"
                                    },
                                    "password":{
                                        "type":"string",
                                        "required":"true", 
                                        "example":"demoPassword"
                                    }
                                }          
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content":{
                            "application/json":{
                                "schema": {
                                    "type": "object",
                                    "properties":{
                                        "access_token":{
                                            "type":"string",
                                            "example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                                        },
                                        "token_type":{
                                            "type":"string",
                                            "example":"Bearer"
                                        },
                                        "expires_in":{
                                            "type":"intager",
                                            "example":"86400"
                                        }
                                    }
                                },
                                "example":{
                                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                                    "token_type": "Bearer",
                                    "expires_in": 86400
                                  }
                            }
                        }                   
                    },
                    "400": {
                        "description": "Bad Request",
                        "content":{
                            "application/json":{
                                "schema": {
                                    "type": "string",
                                    "properties":{
                                        "message":{
                                            "type":"string",
                                            "example":"invalid login - user does not exist"
                                        }
                                    }
                                },
                                "example":{
                                    "message": "invalid login - user does not exist"
                                  }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content":{
                            "application/json":{
                                "schema": {
                                    "type": "string",
                                    "properties":{
                                        "message":{
                                            "type":"string",
                                            "example":"invalid login - bad password"
                                        }
                                    }
                                },
                                "example":{
                                    "message": "invalid login - bad password"
                                  }
                            }
                        }
                    }
                }       
            }
        },
...
    }
} 